+++ src/gamecube_adapter.c
@@ -7,6 +7,7 @@
  */
 
 #include "gamecube_adapter.h"
+#include "usb_device_map.h"
 #include "config.h"
 #include "tusb.h"
 #include "ssd1306.h"
@@ -408,6 +409,7 @@ void gc_mount_cb(uint8_t dev_addr) {
     
     gc_adapter_t* adapter = allocate_adapter(dev_addr);
     if (adapter) {
+        usb_map_register_gamepad(dev_addr, "GameCube");
 #if ENABLE_SERIAL_LOGGING
         printf("GC: Adapter registered!\n");
         printf("GC: Sending initialization command to instance 0...\n");
@@ -444,6 +446,7 @@ void gc_unmount_cb(uint8_t dev_addr) {
 #if ENABLE_SERIAL_LOGGING
     printf("GC: Adapter unmounted at address %d\n", dev_addr);
 #endif
+    usb_map_unregister_gamepad(dev_addr);
     free_adapter(dev_addr);
 }