+++ src/ps3_controller.c
@@ -8,6 +8,7 @@
 
 #include "ps3_controller.h"
 #include "mount_splash.h"
+#include "usb_device_map.h"
 #include "config.h"
 #include "tusb.h"
 #include "ssd1306.h"
@@ -320,6 +321,7 @@ void ps3_mount_cb(uint8_t dev_addr) {
     ps3_controller_t* ctrl = allocate_controller(dev_addr);
     if (ctrl) {
         printf("PS3: Controller registered!\n");
+        usb_map_register_gamepad(dev_addr, "PS3");
         
         // PS3 DualShock 3 requires special initialization
         // Send Feature Report 0xF4 to enable the controller
@@ -350,6 +352,7 @@ void ps3_mount_cb(uint8_t dev_addr) {
 
 void ps3_unmount_cb(uint8_t dev_addr) {
     printf("PS3: Controller unmounted at address %d\n", dev_addr);
+    usb_map_unregister_gamepad(dev_addr);
     free_controller(dev_addr);
 }